requirements.txt
PennyLane == 0.29.0
einops == 0.6.0
kiwisolver == 1.4.4
matplotlib == 3.7.2
numpy == 1.23.5
pandas == 1.5.1
piqa == 1.3.1
qW_Map == 0.1.2
qiskit_aer == 0.11.2
qiskit_machine_learning == 0.5.0
qiskit_terra == 0.24.0
ray == 2.4.0
scikit_learn == 1.1.3
seaborn == 0.12.2
tensorboard == 2.13.0
torch == 1.13.1
torchmetrics == 0.11.4
torchvision == 0.14.1
tqdm == 4.64.1
